+++ src/main/cenario3.cpp
@@ -165,7 +165,7 @@ void fluxoGnomeSort(fstream &entrada, fstream &dataset, fstream &saida, int vers
         conjuntos[i] = leituraDadosVetor(dataset, n); // Colocando dados no conjunto
 
         // Métricas de desempenho
-        int numComparacoes = 0, numCopias = 0;
+        unsigned long long int numComparacoes = 0, numCopias = 0;
 
         // Ponto de inicio de contagem para tempo de execução do algoritmo
         auto inicio = chrono::high_resolution_clock::now();
